# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Formly - Online Form Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /forms/{id}/responses:
    get:
      operationId: listResponses
      tags: [responses]
      summary: List responses for a form
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated responses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponses"
    post:
      operationId: submitResponse
      tags: [responses]
      summary: Submit a form response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitResponseBody"
      responses:
        "201":
          description: Submitted response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /forms/{id}/responses/export:
    get:
      operationId: exportResponsesCsv
      tags: [responses]
      summary: Export responses for a form as CSV
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
  /forms/{formId}/responses/{responseId}:
    get:
      operationId: getResponse
      tags: [responses]
      summary: Get a specific response
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
        - name: responseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Response with answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithAnswers"
...
  /forms/{id}/sheets:
    get:
      operationId: getSheetIntegration
      tags: [sheets]
      summary: Get Google Sheets integration for a form
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sheet integration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetIntegration"
    post:
      operationId: saveSheetIntegration
      tags: [sheets]
      summary: Save/update Google Sheets integration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSheetIntegrationBody"
      responses:
        "200":
          description: Saved integration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetIntegration"
    delete:
      operationId: deleteSheetIntegration
      tags: [sheets]
      summary: Remove Google Sheets integration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /forms/{id}/sheets/sync:
    post:
      operationId: syncToSheets
      tags: [sheets]
      summary: Sync form responses to Google Sheets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                type: object
                properties:
                  rowsSynced:
                    type: integer
                  message:
                    type: string
                required: [rowsSynced, message]
components:
  schemas:
...
    Response:
      type: object
      properties:
        id:
          type: string
        formId:
          type: string
        completed:
          type: boolean
        submittedAt:
          type: string
          format: date-time
      required: [id, formId, completed, submittedAt]
    ResponseWithAnswers:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            answers:
              type: array
              items:
                $ref: "#/components/schemas/Answer"
          required: [answers]
    Answer:
      type: object
      properties:
        id:
          type: string
        questionId:
          type: string
        questionTitle:
          type: string
        value:
          type: string
          nullable: true
      required: [id, questionId, questionTitle]
    SubmitResponseBody:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: string
              value:
                type: string
            required: [questionId]
        completed:
          type: boolean
          default: true
      required: [answers]
    PaginatedResponses:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ResponseWithAnswers"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
...
    SheetIntegration:
      type: object
      properties:
        id:
          type: string
        formId:
          type: string
        spreadsheetId:
          type: string
        spreadsheetName:
          type: string
        sheetName:
          type: string
        enabled:
          type: boolean
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, formId, spreadsheetId, spreadsheetName, sheetName, enabled, createdAt]
    SaveSheetIntegrationBody:
      type: object
      properties:
        spreadsheetId:
          type: string
        spreadsheetName:
          type: string
        sheetName:
          type: string
        enabled:
          type: boolean
          default: true
      required: [spreadsheetId, spreadsheetName, sheetName]